requirements.txt
streamlit
Pillow
pandas
numpy
google-generativeai
requests